+++ packages/media/src/room/eventContainer.ts
@@ -1,21 +1,32 @@
-import { Participant, Room, RoomEvent } from 'livekit-client';
-import { sequenceHandler } from '../helper';
-import { TargetParticipant, TargetParticipantFactory } from '../participant/targetParticipant';
+import { Participant, Room, RoomEvent } from "livekit-client";
+import { sequenceHandler } from "../helper";
+import {
+  TargetParticipant,
+  TargetParticipantFactory,
+} from "../participant/targetParticipant";
 import {
   ConnectionState,
   CurrentConnectionInfo,
   OnConnectionStateChanged,
   OnParticipantConnected,
   OnParticipantDisconnected,
+  OnPresenterUpdated,
   RoomHandler,
-} from '../types';
+} from "../types";
+import DataChannelContainer, { MetadataSort } from "./dataChannelContainer";
 
 class RoomEventContainer {
   #room: Room;
   #handler: RoomHandler;
+  #dataChannelContainer: DataChannelContainer;
 
-  constructor(room: Room, handler: RoomHandler) {
+  constructor(
+    room: Room,
+    dataChannelContainer: DataChannelContainer,
+    handler: RoomHandler
+  ) {
     this.#room = room;
+    this.#dataChannelContainer = dataChannelContainer;
     this.#handler = handler;
   }
 
@@ -23,7 +34,8 @@ class RoomEventContainer {
     this.#room
       .on(RoomEvent.ParticipantConnected, this.onParticipantConnected)
       .on(RoomEvent.ParticipantDisconnected, this.onParticipantDisconnected)
-      .on(RoomEvent.ConnectionStateChanged, this.onConnectionStateChanged);
+      .on(RoomEvent.ConnectionStateChanged, this.onConnectionStateChanged)
+      .on(RoomEvent.RoomMetadataChanged, this.onRoomMetadataChanged);
   };
 
   initializeCurrentRoomStatus = (participant: Participant) => {
@@ -31,19 +43,25 @@ class RoomEventContainer {
   };
 
   onParticipantConnected = async (participant: Participant) => {
-    const targetParticipant = await TargetParticipantFactory.create(participant);
+    const targetParticipant = await TargetParticipantFactory.create(
+      participant
+    );
 
-    sequenceHandler<OnParticipantConnected, [TargetParticipant]>(this.#handler.onParticipantConnected, [
-      targetParticipant,
-    ]);
+    sequenceHandler<OnParticipantConnected, [TargetParticipant]>(
+      this.#handler.onParticipantConnected,
+      [targetParticipant]
+    );
   };
 
   onParticipantDisconnected = async (participant: Participant) => {
-    const targetParticipant = await TargetParticipantFactory.create(participant);
+    const targetParticipant = await TargetParticipantFactory.create(
+      participant
+    );
 
-    sequenceHandler<OnParticipantDisconnected, [TargetParticipant]>(this.#handler.onParticipantDisconnected, [
-      targetParticipant,
-    ]);
+    sequenceHandler<OnParticipantDisconnected, [TargetParticipant]>(
+      this.#handler.onParticipantDisconnected,
+      [targetParticipant]
+    );
   };
 
   onConnectionStateChanged = async (connectionState: ConnectionState) => {
@@ -52,9 +70,25 @@ class RoomEventContainer {
       status: connectionState,
     };
 
-    sequenceHandler<OnConnectionStateChanged, [CurrentConnectionInfo]>(this.#handler.onConnectionStateChanged, [
-      currentConnectionInfo,
-    ]);
+    sequenceHandler<OnConnectionStateChanged, [CurrentConnectionInfo]>(
+      this.#handler.onConnectionStateChanged,
+      [currentConnectionInfo]
+    );
+  };
+
+  onRoomMetadataChanged = async (metadata: string) => {
+    const data = this.#dataChannelContainer.onFetchMetaDataHandler(metadata);
+
+    if (data.type === MetadataSort.PRESENTER) {
+      const targetParticipant = await TargetParticipantFactory.create(
+        data.payload as Participant
+      );
+      console.log(targetParticipant);
+      sequenceHandler<OnPresenterUpdated, [TargetParticipant]>(
+        this.#handler.onPresenterUpdated,
+        [targetParticipant]
+      );
+    }
   };
 }